#19 ishu/anjum
What changes have you made?
Added a function that will ask if the opponent has arrived or not.
Function that will display the opponent arrival status.
*

Why do we need this change?
This change will update the host player of the opponent arrival so that the game can start.
*

What can this change impact?
This change will keep the host player update of the opponent's arrival status.